Refactored loose coupling components and GraphQL queries
This commit introduces a degree of refactoring to the loose coupling components by creating new files: 'useLooseCoupledForm.ts', 'useLooseCoupledView.ts' and 'useLooseCoupledData.ts'. This helps to create a more organized and clearer structure. In addition, there were updates applied to the GraphQL queries which included the creation of the 'createIssueMutation' for better data management.